AJAX functionality to get new data
Moved all behavior needed to start the map into MapBox's constructor.  Added new
method get_recent_purchases, which makes an AJAX call to get new data back in JSON
format, which is then added to the list of markers.  For now, this is set to be called
every 5 seconds, so markers keep populating the map.