Controller and MapBox object pass time of last query back and forth
Updated controller to send the time of the query back to the AJAX call, which is then set
to variable @last_query_time.  Subsequent AJAX calls from MapBox include this time as data
for the controller to use when deciding which purchases to pull, and how long to set the
delay time for each marker.