Sort markers by display time
Sorted the markers to be added to the map, in preparation of eventually staggering the
addition of markers from one big addition of all markers of each call, to be incrementally
added over time.  This will reduce the cycles needed to loop through each marker on each
frame of the animation.