Markers are added to the map in chunks
Previously after a single AJAX call for marker data, all the markers would be added to the
map that were received in one big chunk.  All these markers are then looped over in the
update loop each frame, even though for big groups of markers, most of them are not to
be displayed yet.  So instead the received markers are added to the map in chunks of, for
now, 10.