Marker constructor arguments changed
Marker class used to accept each parameter individually, such as "id", "lat", "radius",
but now it just accepts the JSON data for the marker itself.  Gradually over time the number
of parameters will grow, so to simplify the constructor call, it now just accepts the actual
JSON object itself, and gets the parameters out of it itself.