New marker for error messages
Added a new marker, ErrorMarker, which doesn't fade out.  Instead it has a blinking red dot
in the center.  Clicking it will bring up an InfoWindow.

Also moved the child classes of Marker to a separate folder, so that they are loaded after Marker
is in the assets pipeline properly.