Basic geocoding support
Added the ability to do geocoding for the Purchase model.  As is, I can't easily generate
dummy addresses, and there are limits to the number of API calls for geocoding, so I am
leaving this as just random lat/lng for now.  To change this, uncommenting the
"after_validation :geocode" part will do the trick.